add transformer capability to custom bases
